@@ -1,6 +1,9 @@
 {
   "name": "angular-auth-oidc-client",
   "version": "0.0.0",
+  "engines": {
+    "node": "^18.0.0"
+  },
   "private": true,
   "scripts": {
     "docusaurus": "docusaurus",
@@ -14,14 +17,14 @@
     "write-heading-ids": "docusaurus write-heading-ids"
   },
   "dependencies": {
-    "@docusaurus/core": "2.3.1",
-    "@docusaurus/preset-classic": "2.3.1",
-    "@mdx-js/react": "^1.6.21",
+    "@docusaurus/core": "^3.0.0",
+    "@docusaurus/preset-classic": "^3.0.0",
+    "@docusaurus/remark-plugin-npm2yarn": "^3.0.0",
     "@svgr/webpack": "^5.5.0",
     "clsx": "^1.1.1",
     "file-loader": "^6.2.0",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "url-loader": "^4.1.1"
   },
   "browserslist": {
