@@ -53,39 +53,39 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^16.0.2",
-    "@angular/common": "^16.0.2",
-    "@angular/compiler": "^16.0.2",
-    "@angular/core": "^16.0.2",
-    "@angular/forms": "^16.0.2",
-    "@angular/platform-browser": "^16.0.2",
-    "@angular/platform-browser-dynamic": "^16.0.2",
-    "@angular/router": "^16.0.2",
+    "@angular/animations": "^17.0.8",
+    "@angular/common": "^17.0.8",
+    "@angular/compiler": "^17.0.8",
+    "@angular/core": "^17.0.8",
+    "@angular/forms": "^17.0.8",
+    "@angular/platform-browser": "^17.0.8",
+    "@angular/platform-browser-dynamic": "^17.0.8",
+    "@angular/router": "^17.0.8",
     "@schematics/angular": "^16.0.2",
     "rxjs": "^7.4.0",
     "shelljs": "^0.8.5",
     "tslib": "^2.1.0",
-    "zone.js": "~0.13.0"
+    "zone.js": "~0.14.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^16.0.2",
-    "@angular-devkit/core": "^16.0.2",
-    "@angular-devkit/schematics": "^16.0.2",
-    "@angular-eslint/builder": "^16.0.2",
-    "@angular-eslint/eslint-plugin": "^16.0.2",
-    "@angular-eslint/eslint-plugin-template": "^16.0.2",
-    "@angular-eslint/schematics": "^16.0.2",
-    "@angular-eslint/template-parser": "^16.0.2",
-    "@angular/cli": "^16.0.2",
-    "@angular/compiler-cli": "^16.0.2",
+    "@angular-devkit/build-angular": "^17.0.8",
+    "@angular-devkit/core": "^17.0.8",
+    "@angular-devkit/schematics": "^17.0.8",
+    "@angular-eslint/builder": "^17.1.1",
+    "@angular-eslint/eslint-plugin": "^17.1.1",
+    "@angular-eslint/eslint-plugin-template": "^17.1.1",
+    "@angular-eslint/schematics": "^17.1.1",
+    "@angular-eslint/template-parser": "^17.1.1",
+    "@angular/cli": "^17.0.8",
+    "@angular/compiler-cli": "^17.0.8",
     "@evilmartians/lefthook": "^1.0.3",
     "@types/jasmine": "^4.0.0",
     "@types/node": "^16.11.7",
-    "@typescript-eslint/eslint-plugin": "^5.59.2",
-    "@typescript-eslint/parser": "^5.59.2",
+    "@typescript-eslint/eslint-plugin": "^6.10.0",
+    "@typescript-eslint/parser": "^6.10.0",
     "copyfiles": "^2.4.1",
     "coveralls": "^3.1.0",
-    "eslint": "^8.39.0",
+    "eslint": "^8.53.0",
     "eslint-plugin-import": "2.27.5",
     "eslint-plugin-jsdoc": "44.2.4",
     "eslint-plugin-prefer-arrow": "1.2.3",
@@ -97,10 +97,10 @@
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "^2.0.0",
-    "ng-packagr": "^16.0.1",
+    "ng-packagr": "^17.0.3",
     "prettier": "^2.2.1",
     "rfc4648": "^1.5.0",
     "ts-node": "^10.9.1",
-    "typescript": "5.0.4"
+    "typescript": "~5.2.0"
   }
 }
