@@ -24,6 +24,6 @@ export interface AuthResult {
   // eslint-disable-next-line @typescript-eslint/naming-convention
   session_state?: any;
   state?: any;
-  // eslint-disable-next-line @typescript-eslint/naming-convention
+  scope?: string;
   expires_in?: number;
 }
